+++ x/crosschain/genesis.go
@@ -1,10 +1,8 @@
 package crosschain
 
 import (
-	sdkmath "cosmossdk.io/math"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
-	"github.com/zeta-chain/node/pkg/coin"
 	"github.com/zeta-chain/node/x/crosschain/keeper"
 	"github.com/zeta-chain/node/x/crosschain/types"
 )
@@ -12,9 +10,8 @@ import (
 // InitGenesis initializes the crosschain module's state from a provided genesis
 // state.
 func InitGenesis(ctx sdk.Context, k keeper.Keeper, genState types.GenesisState) {
-	// Always set the zeta accounting to zero at genesis.
-	// ZetaAccounting value is build by iterating through all the cctxs and adding the amount to the zeta accounting.
-	k.SetZetaAccounting(ctx, types.ZetaAccounting{AbortedZetaAmount: sdkmath.ZeroUint()})
+	k.SetZetaAccounting(ctx, genState.ZetaAccounting)
+
 	// Set all the outbound tracker
 	for _, elem := range genState.OutboundTrackerList {
 		k.SetOutboundTracker(ctx, elem)
@@ -37,8 +34,6 @@ func InitGenesis(ctx sdk.Context, k keeper.Keeper, genState types.GenesisState) 
 		}
 	}
 
-	// Set all the chain nonces
-
 	// Set all the last block heights
 	for _, elem := range genState.LastBlockHeightList {
 		if elem != nil {
@@ -46,36 +41,15 @@ func InitGenesis(ctx sdk.Context, k keeper.Keeper, genState types.GenesisState) 
 		}
 	}
 
-	// Set all the cross-chain txs
-	//tss, found := k.GetObserverKeeper().GetTSS(ctx)
-	//if found {
+	// Set the cross-chain transactions only,
+	// We don't need to call SaveCCTXUpdate as the other fields are being set already
 	for _, elem := range genState.CrossChainTxs {
 		if elem != nil {
 			cctx := *elem
 			k.SetCrossChainTx(ctx, cctx)
-			// set mapping inboundHash -> cctxIndex
-			//in, _ := k.GetInboundHashToCctx(ctx, cctx.InboundParams.ObservedHash)
-			//in.InboundHash = cctx.InboundParams.ObservedHash
-			//found := false
-			//for _, cctxIndex := range in.CctxIndex {
-			//	if cctxIndex == cctx.Index {
-			//		found = true
-			//		break
-			//	}
-			//}
-			//if !found {
-			//	in.CctxIndex = append(in.CctxIndex, cctx.Index)
-			//}
-			//k.SetInboundHashToCctx(ctx, in)
-
-			if cctx.CctxStatus.Status == types.CctxStatus_Aborted &&
-				cctx.InboundParams.CoinType == coin.CoinType_Zeta &&
-				cctx.CctxStatus.IsAbortRefunded == false {
-				k.AddZetaAbortedAmount(ctx, keeper.GetAbortedAmount(cctx))
-			}
 		}
 	}
-	//}
+
 	for _, elem := range genState.FinalizedInbounds {
 		k.SetFinalizedInbound(ctx, elem)
 	}